+++ node/src/main/kotlin/net/corda/node/internal/NodeStartup.kt
@@ -17,6 +17,7 @@ import net.corda.node.shell.InteractiveShell
 import net.corda.node.utilities.registration.HTTPNetworkRegistrationService
 import net.corda.node.utilities.registration.NetworkRegistrationHelper
 import net.corda.nodeapi.internal.addShutdownHook
+import net.corda.nodeapi.internal.config.UnknownConfigurationKeysException
 import org.fusesource.jansi.Ansi
 import org.fusesource.jansi.AnsiConsole
 import org.slf4j.bridge.SLF4JBridgeHandler
@@ -86,6 +87,9 @@ open class NodeStartup(val args: Array<String>) {
             } else {
                 conf0
             }
+        } catch (e: UnknownConfigurationKeysException) {
+            logger.error(e.message)
+            return false
         } catch (e: Exception) {
             logger.error("Exception during node configuration", e)
             return false